Christina

See merge request compsci308_2019spring/slogo_team02!177

+++ src/Backend/src/Models/TurtleModel.java
@@ -80,7 +80,7 @@ public class TurtleModel {
     }
 
     public Turtle getCurrentTurtle() {
-        return allTurtles.get(currentTurtleIndex);
+        return allTurtles.get((int) currentTurtleIndex);
     }
 
     public Turtle getTurtle(int turtleId) {
